Implement IdentifierContainerInterface on Assertion
