+++ src/IdentifierContainerInterface.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace webignition\BasilModel\Action;
+namespace webignition\BasilModel;
 
 use webignition\BasilModel\Identifier\IdentifierInterface;
 
-interface IdentifierContainerInterface extends ActionInterface
+interface IdentifierContainerInterface
 {
     public function getIdentifier(): ?IdentifierInterface;
     public function withIdentifier(IdentifierInterface $identifier): IdentifierContainerInterface;
